requirements.txt
django==3.2.0
djangorestframework==3.13.1
gunicorn==20.1.0